pyproject.toml
[project]
name = "research-bot"
version = "0.1.0"
description = "AI-powered research assistant using CrewAI"
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai>=0.80.0",
    "crewai-tools>=0.14.0",
    "tavily-python>=0.5.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.3.0",
]

[project.scripts]
research-bot = "research_bot.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/research_bot"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
